Refactor Cosdata SDK into submodules
