+++ src/cosdata/client.py
@@ -163,7 +163,7 @@ class Client:
         self, 
         collection_name: str,
         distance_metric: str = "cosine"
-    ) -> Dict[str, Any]:
+    ) -> Index:
         """
         Create an index for the specified collection.
         
@@ -173,85 +173,7 @@ class Client:
         Returns:
             JSON response from the server
         """
-        return Index(self, collection_name).create(distance_metric)
-    
-    def create_transaction(self, collection_name: str) -> 'Transaction':
-        """
-        Create a new transaction for the specified collection.
+        index = Index(self, collection_name)
+        index.create(distance_metric)
         
-        Args:
-            collection_name: Name of the collection
-            
-        Returns:
-            Transaction object
-        """
-        return Transaction(self, collection_name)
-    
-    def search_vector(
-        self, 
-        collection_name: str, 
-        vector: List[float], 
-        nn_count: int = 5
-    ) -> Dict[str, Any]:
-        """
-        Search for nearest neighbors of a vector.
-        
-        Args:
-            collection_name: Name of the collection to search in
-            vector: Vector to search for similar vectors
-            nn_count: Number of nearest neighbors to return
-            
-        Returns:
-            Search results
-        """
-        url = f"{self.base_url}/search"
-        data = {
-            "vector_db_name": collection_name,
-            "vector": vector,
-            "nn_count": nn_count
-        }
-        
-        response = requests.post(
-            url, 
-            headers=self._get_headers(), 
-            data=json.dumps(data), 
-            verify=self.verify_ssl
-        )
-        
-        if response.status_code != 200:
-            raise Exception(f"Failed to search vector: {response.text}")
-            
-        return response.json()
-    
-    def fetch_vector(
-        self, 
-        collection_name: str, 
-        vector_id: Union[str, int]
-    ) -> Dict[str, Any]:
-        """
-        Fetch a specific vector by ID.
-        
-        Args:
-            collection_name: Name of the collection
-            vector_id: ID of the vector to fetch
-            
-        Returns:
-            Vector data
-        """   
-        url = f"{self.base_url}/fetch"
-        data = {
-            "vector_db_name": collection_name,
-            "vector_id": vector_id
-        }
-        
-        response = requests.post(
-            url, 
-            headers=self._get_headers(), 
-            data=json.dumps(data), 
-            verify=self.verify_ssl
-        )
-        
-        if response.status_code != 200:
-            raise Exception(f"Failed to fetch vector: {response.text}")
-            
-        return response.json()
+        return index
